lobby_list_update message implementation
Took some wrangling to get this message class implemented. I couldn't send Lobby objects through the socket - I was trying to JSON serialize the message instead of the Lobby (like an idiot) Now the server stores a list of lobbies as a dict and sends that dict to the clients. Will implement a separate list to hold the actual Lobby objects shortly.