Click to join lobbies
Users can now click on the lobby names and join them. Currently no way to enter passwords, so all are set to empty strings. Right now it's deconstructing and reconstructing the Lobby object when sending from the server to each client instance. That seems fine for now but may want to revisit that later? Or just triple check the JSON serialization/deserialization process